refactor: cleaner interface to averaging routines
