+++ tests/test_averaging.py
@@ -5,12 +5,16 @@ import pytest
 from pytest_cases import fixture_ref as fxref
 from pytest_cases import parametrize_plus
 
+import src.edges_analysis.analysis.averaging
+
 
 def test_non_stationary_avg_1d_shape():
     x = np.linspace(0, 1, 20)
     y = x ** 2
 
-    avg = tools.non_stationary_weighted_average(data=y, x=x, n_terms=3)
+    avg = src.edges_analysis.analysis.averaging.non_stationary_weighted_average(
+        data=y, x=x, n_terms=3
+    )
 
     assert isinstance(avg, float)
     assert np.isclose(avg, np.mean(y))
@@ -20,7 +24,9 @@ def test_non_stationary_avg_2d_shape():
     x = np.linspace(0, 1, 20)
     y = np.array([x ** 2 for _ in range(4)])
 
-    avg = tools.non_stationary_weighted_average(data=y, x=x, n_terms=3)
+    avg = src.edges_analysis.analysis.averaging.non_stationary_weighted_average(
+        data=y, x=x, n_terms=3
+    )
 
     assert avg.shape == (4,)
     assert np.allclose(avg, np.mean(y))
@@ -33,7 +39,9 @@ def test_non_stationary_avg_1d_weights():
     weights = np.ones(len(x), dtype=bool)
     weights[::4] = 0
 
-    avg = tools.non_stationary_weighted_average(data=y, x=x, n_terms=3, weights=weights)
+    avg = src.edges_analysis.analysis.averaging.non_stationary_weighted_average(
+        data=y, x=x, n_terms=3, weights=weights
+    )
     assert isinstance(avg, float)
     assert np.isclose(avg, np.mean(y))
 
@@ -45,7 +53,9 @@ def test_non_stationary_avg_1d_nonzero_weights():
     weights = np.linspace(1, 4, len(x))
     #    weights[::4] = 0
 
-    avg = tools.non_stationary_weighted_average(data=y, x=x, n_terms=3, weights=weights)
+    avg = src.edges_analysis.analysis.averaging.non_stationary_weighted_average(
+        data=y, x=x, n_terms=3, weights=weights
+    )
     assert isinstance(avg, float)
     assert np.isclose(avg, np.mean(y))
 
@@ -59,7 +69,7 @@ def test_non_stationary_avg_1d_with_model_fit():
 
     weights = np.linspace(1, 4, len(x))
 
-    avg = tools.non_stationary_weighted_average(
+    avg = src.edges_analysis.analysis.averaging.non_stationary_weighted_average(
         data=y, x=x, n_terms=3, weights=weights, model_fit=model
     )
     assert isinstance(avg, float)
@@ -74,7 +84,7 @@ def test_non_stationary_avg_1d_with_model():
 
     weights = np.linspace(1, 4, len(x))
 
-    avg = tools.non_stationary_weighted_average(
+    avg = src.edges_analysis.analysis.averaging.non_stationary_weighted_average(
         data=y, x=x, n_terms=3, weights=weights, model=model
     )
     assert isinstance(avg, float)
@@ -86,7 +96,7 @@ def test_get_binned_weights_1d():
     x = np.linspace(0, 1, 20)
     bins = [-0.1, 1.1]
 
-    wghts = tools.get_binned_weights(x=x, bins=bins, weights=w)
+    wghts = src.edges_analysis.analysis.averaging.get_binned_weights(x=x, bins=bins, weights=w)
     assert wghts == 20
 
 
@@ -95,7 +105,7 @@ def test_get_binned_weights_2d():
     x = np.linspace(0, 1, 20)
     bins = [-0.1, 1.1]
 
-    wghts = tools.get_binned_weights(x=x, bins=bins, weights=w)
+    wghts = src.edges_analysis.analysis.averaging.get_binned_weights(x=x, bins=bins, weights=w)
     assert wghts.shape == (10, 1)
     assert all(w == 20 for w in wghts)
 
@@ -105,7 +115,7 @@ def test_get_binned_weights_more_bins():
     x = np.linspace(0, 1, 20, endpoint=False)
     bins = np.linspace(0, 1, 11)
 
-    wghts = tools.get_binned_weights(x=x, bins=bins, weights=w)
+    wghts = src.edges_analysis.analysis.averaging.get_binned_weights(x=x, bins=bins, weights=w)
     assert wghts.shape == (10, 10)
     assert np.allclose(wghts, 2)
 
@@ -177,7 +187,7 @@ def test_model_bin_gha(params, weights, refit):
         fit_params = params
         resids = data - model
 
-    p, r, w = tools.model_bin_gha(
+    p, r, w = src.edges_analysis.analysis.averaging.bin_gha_unbiased_regular(
         params=fit_params,
         resids=resids,
         weights=weights,
@@ -206,13 +216,19 @@ class TestBinArray:
     def test_out_shape_no_coords(self, fid_params, ideal_weights):
         model, corrupt, noise = make_data(fid_params, ideal_weights)
 
-        coords, mean, wght = tools.bin_array(corrupt, axis=0)
+        coords, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, axis=0
+        )
         assert mean.shape == coords.shape == wght.shape == (1, 500)
 
-        coords, mean, wght = tools.bin_array(corrupt, axis=1)
+        coords, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, axis=1
+        )
         assert mean.shape == coords.shape == wght.shape == (50, 1)
 
-        coords, mean, wght = tools.bin_array(corrupt, axis=-1)
+        coords, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, axis=-1
+        )
         assert mean.shape == coords.shape == wght.shape == (50, 1)
 
     def test_out_shape_with_coords(self, fid_params, ideal_weights):
@@ -220,37 +236,55 @@ class TestBinArray:
 
         coords = FREQ
         with pytest.raises(ValueError):
-            tools.bin_array(corrupt, coords=coords, axis=0)
+            src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+                corrupt, coords=coords, axis=0
+            )
 
-        outc, mean, wght = tools.bin_array(corrupt, coords=coords, axis=-1)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, coords=coords, axis=-1
+        )
         assert mean.shape == outc.shape == wght.shape == (50, 1)
 
-        outc, mean, wght = tools.bin_array(corrupt, coords=coords, axis=-1, bins=5)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, coords=coords, axis=-1, bins=5
+        )
         assert mean.shape == outc.shape == wght.shape == (50, 100)
 
-        outc, mean, wght = tools.bin_array(corrupt, coords=coords, axis=-1, bins=11)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, coords=coords, axis=-1, bins=11
+        )
         assert mean.shape == outc.shape == wght.shape == (50, 46)
 
-        outc, mean, wght = tools.bin_array(corrupt, coords=coords, axis=-1, bins=1.0)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, coords=coords, axis=-1, bins=1.0
+        )
         assert mean.shape == outc.shape == wght.shape == (50, 50)
 
     def test_shape_3d_input(self, fid_params, ideal_weights):
         model, corrupt, noise = make_data(fid_params, ideal_weights)
 
         corrupt = corrupt.reshape((50, 20, 25))
-        outc, mean, wght = tools.bin_array(corrupt, axis=0, bins=5)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, axis=0, bins=5
+        )
         assert outc.shape == mean.shape == wght.shape == (10, 20, 25)
 
-        outc, mean, wght = tools.bin_array(corrupt, axis=1, bins=5)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, axis=1, bins=5
+        )
         assert outc.shape == mean.shape == wght.shape == (50, 4, 25)
 
-        outc, mean, wght = tools.bin_array(corrupt, axis=2, bins=5)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            corrupt, axis=2, bins=5
+        )
         assert outc.shape == mean.shape == wght.shape == (50, 20, 5)
 
     def test_unity_input(self):
         data = np.ones((10, 100))
 
-        outc, mean, wght = tools.bin_array(data, axis=0, bins=5)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            data, axis=0, bins=5
+        )
         assert np.all(mean == 1)
         assert np.all(wght == 5)
 
@@ -259,6 +293,8 @@ class TestBinArray:
         weights = np.ones((10, 100))
         weights[:, ::5] = 0
 
-        outc, mean, wght = tools.bin_array(data, axis=1, weights=weights, bins=5)
+        outc, mean, wght = src.edges_analysis.analysis.averaging.bin_array_unbiased_irregular(
+            data, axis=1, weights=weights, bins=5
+        )
         assert np.all(mean == 1)
         assert np.all(wght == 4)
